moving old methods to deprecated
